refactor(e2e): extract karpenter CP upgrade test to standalone TestKarpenterUpgradeControlPlane
The control plane upgrade + drift test is the most expensive subtest (~25 min).
Extract it to its own top-level test in karpenter_control_plane_upgrade_test.go
so it can be independently targeted by CI regex.

Since TestKarpenter no longer performs an upgrade:
- Remove PreviousReleaseImage and WithUpgradeTarget
- Drop HighlyAvailable control plane policy (not needed for provisioning tests)

Made-with: Cursor
